+++ src/new_destiny/__init__.py
@@ -1,0 +1,29 @@
+from .json_types import (
+    JSONArray,
+    JSONObject,
+    JSONPrimitive,
+    JSONValue,
+    RiotResponse,
+    expect_array,
+    expect_bool,
+    expect_float,
+    expect_int,
+    expect_number,
+    expect_object,
+    expect_string,
+)
+
+__all__ = [
+    "JSONArray",
+    "JSONObject",
+    "JSONPrimitive",
+    "JSONValue",
+    "RiotResponse",
+    "expect_array",
+    "expect_bool",
+    "expect_float",
+    "expect_int",
+    "expect_number",
+    "expect_object",
+    "expect_string",
+]